Refactor:renderContent function is updated
