+++ ConcordiaMaps/components/NextEventModal.js
@@ -1,21 +1,15 @@
 import React, { useEffect, useState } from "react";
-import {
-  View,
-  Text,
-  Modal,
-  TouchableOpacity,
-  ActivityIndicator,
-} from "react-native";
+import { View, Text, Modal } from "react-native";
 import * as Calendar from "expo-calendar";
-import { format } from "date-fns";
 import styles from "../styles";
 import PropTypes from "prop-types";
 import { useNavigation } from "@react-navigation/native";
+import { EventLoading, EventDetails, NoEvents } from "./EventComponents"; // Adjust path as needed
 
 const NextEventModal = ({ isVisible, onClose }) => {
   const [nextEvent, setNextEvent] = useState(null);
   const [loading, setLoading] = useState(true);
-  const [timeRemaining, setTimeRemaining] = useState(0); // Store the remaining time in seconds
+  const [timeRemaining, setTimeRemaining] = useState(0);
   const navigation = useNavigation();
 
   useEffect(() => {
@@ -26,13 +20,11 @@ const NextEventModal = ({ isVisible, onClose }) => {
 
   useEffect(() => {
     let intervalId;
-
     if (nextEvent && timeRemaining > 0) {
       intervalId = setInterval(() => {
         setTimeRemaining((prevTime) => prevTime - 1);
       }, 1000);
     }
-
     return () => clearInterval(intervalId);
   }, [nextEvent, timeRemaining]);
 
@@ -101,26 +93,13 @@ const NextEventModal = ({ isVisible, onClose }) => {
     }
   }, [nextEvent]);
 
-  const formatTime = (seconds) => {
-    const minutes = Math.floor(seconds / 60);
-    const remainingSeconds = seconds % 60;
-    return `${String(minutes).padStart(2, "0")}:${String(remainingSeconds).padStart(2, "0")}`;
-  };
-
   const handleGetDirections = async () => {
-    // Set the default origin coordinates
     const defaultOrigin = {
       latitude: 45.494971642137095,
       longitude: -73.57791280320929,
     };
-
-    // For the destination, use the event's location if available; otherwise, use null
     const destination = nextEvent.location;
-
-    // Close the modal before navigating
     onClose();
-
-    // Navigate to the GetDirections screen with the origin and destination
     navigation.navigate("GetDirections", {
       origin: defaultOrigin,
       destination,
@@ -129,62 +108,17 @@ const NextEventModal = ({ isVisible, onClose }) => {
   };
 
   const renderContent = () => {
-    if (loading) {
-      return <ActivityIndicator size="large" color="#912338" />;
-    }
-    if (nextEvent) {
+    if (loading) return <EventLoading />;
+    if (nextEvent)
       return (
-        <>
-          <View style={styles.eventContainer}>
-            <View style={styles.timeCircle}>
-              <Text style={styles.timeText}>{formatTime(timeRemaining)}</Text>
-            </View>
-            <View style={styles.eventDetails}>
-              <Text style={styles.eventTitleNext}>{nextEvent.title}</Text>
-              <Text style={styles.eventInfoNext}>
-                {format(new Date(nextEvent.startDate), "hh:mm a")} -{" "}
-                {format(new Date(nextEvent.endDate), "hh:mm a")}
-              </Text>
-              {nextEvent.location && (
-                <Text style={[styles.eventInfoNext, { marginBottom: 5 }]}>
-                  {nextEvent.location}
-                </Text>
-              )}
-            </View>
-          </View>
-          <View
-            style={[
-              styles.toggleContainer,
-              {
-                flexDirection: "row",
-                justifyContent: "space-between",
-                marginTop: -1,
-              },
-            ]}
-          >
-            <TouchableOpacity
-              onPress={handleGetDirections}
-              style={[styles.closeButton, { marginRight: 10 }]}
-            >
-              <Text style={styles.closeButtonText}>Get Directions</Text>
-            </TouchableOpacity>
-            <TouchableOpacity onPress={onClose} style={styles.closeButton}>
-              <Text style={styles.closeButtonText}>Close</Text>
-            </TouchableOpacity>
-          </View>
-        </>
+        <EventDetails
+          event={nextEvent}
+          timeRemaining={timeRemaining}
+          onGetDirections={handleGetDirections}
+          onClose={onClose}
+        />
       );
-    }
-    return (
-      <View style={{ alignItems: "center", marginTop: 10 }}>
-        <Text style={{ fontSize: 16, marginBottom: 10 }}>
-          No upcoming events today.
-        </Text>
-        <TouchableOpacity onPress={onClose} style={styles.closeButton}>
-          <Text style={styles.closeButtonText}>Close</Text>
-        </TouchableOpacity>
-      </View>
-    );
+    return <NoEvents onClose={onClose} />;
   };
 
   return (
